+++ ogre/src/OgreRayQuery.cc
@@ -167,7 +167,7 @@ RayQueryResult OgreRayQuery::ClosestPoint(bool /*_forceSceneUpdate*/) // NOLINT
             OgreConversions::Convert(
             ogreEntity->getParentNode()->_getDerivedScale()));
 
-        for (unsigned int i = 0; i < indexCount; i += 3)
+        for (size_t i = 0; i < indexCount; i += 3)
         {
           // when indices size is not divisible by 3
           if (i+2 >= indexCount)
@@ -221,7 +221,7 @@ void OgreRayQuery::MeshInformation(const Ogre::Mesh *_mesh,
   _vertex_count = _index_count = 0;
 
   // Calculate how many vertices and indices we're going to need
-  for (uint16_t i = 0; i < _mesh->getNumSubMeshes(); ++i)
+  for (size_t i = 0; i < _mesh->getNumSubMeshes(); ++i)
   {
     Ogre::SubMesh* submesh = _mesh->getSubMesh(i);
 
@@ -250,7 +250,7 @@ void OgreRayQuery::MeshInformation(const Ogre::Mesh *_mesh,
   added_shared = false;
 
   // Run through the submeshes again, adding the data into the arrays
-  for (uint16_t i = 0; i < _mesh->getNumSubMeshes(); ++i)
+  for (size_t i = 0; i < _mesh->getNumSubMeshes(); ++i)
   {
     Ogre::SubMesh* submesh = _mesh->getSubMesh(i);
